% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sampling_points.R
\name{get_sampling_points}
\alias{get_sampling_points}
\title{Função para obter a tabela de pontos por ugrhi}
\usage{
get_sampling_points(n_ugrhi)
}
\arguments{
\item{n_ugrhi}{}
}
\value{

}
\description{
Função para obter a tabela de pontos por ugrhi
}
\examples{
get_sampling_points(6)
}
