% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login_infoaguas.R
\name{login_infoaguas}
\alias{login_infoaguas}
\title{Função para autenticar no Sistema InfoAguas da CETESB}
\usage{
login_infoaguas(login, password)
}
\arguments{
\item{password}{}
}
\value{

}
\description{
Função para autenticar no Sistema InfoAguas da CETESB
}
