% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampling_points}
\alias{sampling_points}
\title{sampling_points}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 635 rows and 9 columns.
}
\usage{
sampling_points
}
\description{
sampling_points
}
\keyword{datasets}
