% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-sampling-points.R
\name{get_sampling_points}
\alias{get_sampling_points}
\title{Fun\u00e7\u00e3o para obter a tabela de pontos de monitoramneto}
\usage{
get_sampling_points()
}
\value{

}
\description{
Fun\u00e7\u00e3o para obter a tabela de pontos de monitoramneto
}
