% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sampling_points.R
\name{get_sampling_points}
\alias{get_sampling_points}
\title{Função para obter a tabela de pontos de monitoramneto}
\usage{
get_sampling_points()
}
\value{

}
\description{
Função para obter a tabela de pontos de monitoramneto
}
