% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login-infoaguas.R
\name{login_infoaguas}
\alias{login_infoaguas}
\title{Fun\u00e7\u00e3o para autenticar no Sistema Infoaguas da CETESB}
\usage{
login_infoaguas(login, password)
}
\arguments{
\item{password}{}
}
\value{

}
\description{
Fun\u00e7\u00e3o para autenticar no Sistema Infoaguas da CETESB
}
